feat(ui): show the product page at / when no accounts exist
Extracts the page into product-page.svelte, rendered by both /product
and the root page when the device has no accounts and no session. The
product page's "Get started" links to /?signin, which bypasses the
product view so the create/import chooser stays reachable.
